pyproject.toml
[project]
name = "slack-digest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-genai>=1.25.0",
]
